+++ frontend/.prettierrc
@@ -1,5 +1,5 @@
 {
-    "singleQuote": true,
+    "singleQuote": false,
     "semi": true,
     "tabWidth": 2,
     "trailingComma": "es5"
