+++ frontend/app/layout.tsx
@@ -14,7 +14,7 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "SkyTrack",
   description: "Generated by create next app",
 };
