+++ frontend/app/login/page.tsx
@@ -1,0 +1,7 @@
+export default function Login() {
+  return (
+    <div>
+      <h1>HI</h1>
+    </div>
+  );
+}
